organized interp code
